Split tarball generation for rpm generation
The offical fedora rpm depends on the url source tarball.
Split tarball generation out so it is easier to generate.

Generalize the naming of the tarball to follow the convention
of <package>-<version>.tar.gz.

Fix opae.spec to use the new source loction opae-<version>/

Generalize the tarball excludes all hidden files

+++ scripts/create-rpms.sh
@@ -28,65 +28,20 @@ mkdir _build
 cd _build
 cmake .. -DOPAE_BUILD_LEGACY=ON  -DOPAE_BUILD_TESTS=ON
 
+# From
+#   Version:        2.0.0
+# To
+#   2.0.0
+version=`grep ^Version "${BUILD_DIR}/../opae.spec" | awk '{ print $2 }'`
 
+${BUILD_DIR}/create-tarball.sh opae-${version}
 cd ../..
-tar --transform='s/opae-sdk/opae/' \
-  --exclude=.git \
-  --exclude=.gitignore \
-  --exclude=.github \
-  --exclude=.travis.yml \
-  --exclude=opae.spec.in \
-  --exclude=opae.spec \
-  --exclude=opae-libs/external \
-  --exclude=opae-libs/plugins/ase \
-  --exclude=opae-libs/cmake/config/libopae-all.spec.in \
-  --exclude=opae-libs/cmake/config/run_coverage_test.sh.in \
-  --exclude=opae-libs/cmake/config/run_coverage_test_local.sh.in\
-  --exclude=external/opae-legacy/tests \
-  --exclude=external/opae-legacy/scripts \
-  --exclude=external/opae-legacy/tools/coreidle \
-  --exclude=external/opae-legacy/tools/hssi \
-  --exclude=platforms \
-  --exclude=samples/base \
-  --exclude=samples/hello_afu \
-  --exclude=samples/dummy_afu \
-  --exclude=samples/hello_mpf_afu \
-  --exclude=samples/intg_xeon_nlb \
-  --exclude=samples/base \
-  --exclude=scripts/build-documentation.sh \
-  --exclude=scripts/build-pypi.sh \
-  --exclude=scripts/build-tests-default.sh \
-  --exclude=scripts/coverage-gtapi-mock-drv.sh  \
-  --exclude=scripts/cover-ase.sh  \
-  --exclude=scripts/cover-py.sh  \
-  --exclude=scripts/cover.sh \
-  --exclude=scripts/create-rpms.sh \
-  --exclude=scripts/docker_rpm.sh \
-  --exclude=scripts/index_generator.py \
-  --exclude=scripts/test-build.sh \
-  --exclude=scripts/test-codingstyle-all.sh \
-  --exclude=scripts/test-codingstyle-cpp.sh \
-  --exclude=scripts/test-gtapi-mock-drv.sh \
-  --exclude=scripts/unit-tests.sh \
-  --exclude=scripts/valgrind \
-  --exclude=scripts/.gitignore \
-  --exclude=scripts/bat \
-  --exclude=scripts/push-documentation.sh \
-  --exclude=tools/extra/ras \
-  --exclude=tools/extra/pac_hssi_config \
-  --exclude=tools/extra/pyfpgadiag \
-  --exclude=tools/extra/pypackager \
-  --exclude=tools/utilities \
-  --exclude=opae-libs/include/opae/cxx/.clang-format \
-  --exclude=opae-libs/libopaecxx/.clang-format \
-  --exclude=opae-libs/pyopae/.clang-format \
-  --exclude=opae-libs/.clang-format \
-  --exclude=opae-libs/.clang-format \
-  -z -c -f opae.tar.gz opae-sdk
 
-
-
-mv opae.tar.gz ~/rpmbuild/SOURCES/
+if [ ! -f opae-${version}.tar.gz ]; then
+    echo "Creating the tarball failed"
+    exit 1
+fi
+mv opae-${version}.tar.gz ~/rpmbuild/SOURCES/
 cp "${BUILD_DIR}/../opae.spec" ~/rpmbuild/SPECS/
 
 echo $"${BUILD_DIR}/../opae.spec"
